Supporting the new Oakton resource model. Closes GH-60
